booth_number_id fix & added pretier and pre commit hook

+++ seeders/20241017130000-role-hierarchy-seed.js
@@ -53,9 +53,6 @@ module.exports = {
 
   down: async (queryInterface) => {
     const roleNames = ROLE_CATALOG.map((role) => role.name);
-    await queryInterface.bulkDelete(
-      "tbl_meta_user_role",
-      { disp_name: roleNames }
-    );
+    await queryInterface.bulkDelete("tbl_meta_user_role", { disp_name: roleNames });
   }
 };
